- Parsing code was moved to it's own package - Outage date converted to time.Time - Outage duration converted to time.Duration - JSON fields renamed to eng
